Ignore incorrect type error in mypy
Mypy thinks the result of the normalization operation is a variable length
tuple: 'tuple[int | float, ...]'. The problem appears to follow from using
a generator to iterate over the original tuple, during which its length
info is lost.

Ignoring the error seems preferable to changing the code to explicitly
check for tuple lengths and handling the cases separately.